LayoutTests/platform/ios-13/imported/w3c/web-platform-tests/xhr/send-entity-body-empty-expected.txt


PASS XMLHttpRequest: send("") - empty entity body (POST)
PASS XMLHttpRequest: send("") - empty entity body (PUT)
FAIL XMLHttpRequest: send("") - empty entity body (HEAD) assert_equals: expected "NO" but got "0"
